+++ backend/app/src/main/java/backend/model/Warrior.java
@@ -29,22 +29,4 @@ public class Warrior extends Unit {
     public void boostAttackSpeed() {
         // TODO: Implement boosting attack speed logic
     }
-
-    @Override
-    public void findTarget(Board board) {
-        int minDist = Integer.MAX_VALUE;
-        // TODO: cycle through castles
-        for (Entry<String, Unit> entry : board.units.entrySet()) {
-            var unit = entry.getValue();
-            if (this.owner == unit.owner)
-                continue;
-            int xdist = Math.abs(this.position.getValue0() - unit.position.getValue0());
-            int ydist = Math.abs(this.position.getValue1() - unit.position.getValue1());
-            int dist = Math.max(xdist, ydist);// https://en.wikipedia.org/wiki/Chebyshev_distance
-            if (dist < minDist) {
-                minDist = dist;
-                target = unit;
-            }
-        }
-    }
 }
