+++ backend/app/src/main/java/backend/model/unit/Unit.java
@@ -1,5 +1,7 @@
 package backend.model.unit;
 
+import java.util.Map.Entry;
+
 import org.javatuples.Pair;
 
 import com.github.shamil.Xid;
@@ -62,8 +64,6 @@ public abstract class Unit {
         this.type = type;
     }
 
-    public abstract void findTarget(Board board);
-
     // Logic under boosting damage for a given unit class
     public abstract void boostDamage();
 
@@ -78,9 +78,29 @@ public abstract class Unit {
         this.target.takeDamage(this.damage);
     }
 
-    
+    // simple targeting scheme used by most units.
+    // units that move differently (like castle that doesn't move at all) should override it
+    public void findTarget(Board board) {
+        int minDist = Integer.MAX_VALUE;
+        // TODO: cycle through castles
+        for (Entry<String, Unit> entry : board.units.entrySet()) {
+            var unit = entry.getValue();
+            if (this.owner == unit.owner)
+                continue;
+            int xdist = Math.abs(this.position.getValue0() - unit.position.getValue0());
+            int ydist = Math.abs(this.position.getValue1() - unit.position.getValue1());
+            int dist = Math.max(xdist, ydist);// https://en.wikipedia.org/wiki/Chebyshev_distance
+            if (dist < minDist) {
+                minDist = dist;
+                target = unit;
+            }
+        }
+    }
+    // simple moving scheme used by most units
+    // units that move differently (like castle that doesn't move at all) should override it
     public void move() {
     	System.out.println("chuj dupa japa pipa");
+    	if(target == null) return;
     	int cur0 = position.getValue0(), cur1 = position.getValue1();
     	int tar0 = target.position.getValue0(), tar1 = target.position.getValue1();
     	
@@ -93,8 +113,7 @@ public abstract class Unit {
     	if(cur0 == tar0 && cur1 == tar0) {
     		// moving would "step onto target". Do nothing
     	} else {
-    		// zaktualizuj fielda pos i przenieś w hashmapie.
+    		position = position.setAt0(cur0).setAt1(cur1);
     	}
     }
 }
-//position = position.setValue0(cur0).setValue1(cur1);
